Reduce overhead of metrics recording
This is an optimization to reduce the amount of objects allocated when the various metrics are recorded, and minimize processing done on the application threads.

The various events are now passed in "raw" form via dedicated objects to the sender thread (rather than first being serialized to a string on the app thread). All of the serialization work is now done on the sender thread, avoiding allocation of any new strings or byte arrays. A single StringBuilder is reused in conjunction with the existing ByteBuffer. The NumberFormat ThreadLocals are no longer needed since they will now only be used by the one thread.

Also included are a couple of thread-safety tweaks to the unit tests.