[build] Remove CircleCI caching
Permission issue break caching in CircleCI. Given that these caches
are set up exactly as documentation expects and are valid for only 15
days, they seem to be too much trouble than what they are worth. This
commit removes the cache for now until we find a better way to handle
this problem.