Don't load unixsocket classes unless used
Shuffle the code in the builder so that we don't load unix socket support
classes unless the client is configured to use unix sockets. Use reflection
in the client to test if the address is a UnixSocketAddress.

This allows the client code to function at runtime in UDP mode without having
jnr-unixsocket artifacts on the classpath.