Add third-party license information
Add a NOTICE file to the binary distribution that details license
situation for each subdirectory containing third-party code. This file
should fullfill our obligations when distributing third-party code in
the jar-with-dependencies package.

Provide a more compact overview in THIRDPARTY.md.

Add a script that was used to generate the two files above.