Add an Automatic-Module-Name entry to the MANIFEST
Ensure that an Automatic-Module-Name entry is added to the final
MANIFEST.MF file in the produced jar. This will allow users of the
library to properly import and use it on the module path with a
stable, well-known name.

This commit does *not* change this library itself to be a modular jar,
nor does it incorporate any features that would break backwards
compatibility with JDK 7-8 clients.