feat: Support Dogstatsd 1.1 protocol for distributions
There is a new experimental Dogstatsd protocol (1.1) which supports sending
multiple values in a single datagram.

This is supported since agent 7.25.0/6.25.0.

This evolution in the protocol allows clients to buffer histogram and
distribution values and send them in fewer payload to the agent (providing a
behavior close to client-side aggregation for those types).

This commit add support for this new protocol for distribution.

Fix #215