feat: Add an API to send multiple samples at once
recordDistributionValues is similar to recordDistributionValue but it lets the
client sends multiple samples in one message using dogstatsd 1.1 protocol.

Because this is a shift compared to how other methods are behaving, these
methods are provided only in DirectStatsDClient which provides direct access to
some low level dogstatsd protocol features.

This is recommended in high performance cases were the overhead of the statsd
library might be significant and the sampling is already done by the client.

Port of https://github.com/DataDog/datadog-go/pull/296.