AMLII-2058 Use correct buffer size for unix sockets
When the client is configured to use UDS via builder's address()
method, the client defaulted to UDP-sized buffer.

This patch adds a method to the ClientChannel to fetch the default
packet size for each transport type, and uses it in the client to
pass the correct value to the processor. The maxPacketSizeBytes field
of the resolved builder is reduced to only storing the user preferred
configuration, if any.