Fix races in the tests
resist_dsd_restart for stream and datagram unix socket contained a
race when testing server closed condition. Sending multiple payloads
in a loop caused a situation when the "Connection refused" error was
triggered twice, confusing the following step of the test that expects
a different error.

jnr-unixsocket does not interrupt receive call when the server channel
is closed, and successfully reads a datagram sent after the call to
close.

This patch modifies the server to actually close the socket when
instructed and send the test payload once to controllably trigger the error.